feat: add event driven score engine

+++ src/goat-shooooting.Framework/RunCompletionTracker.cs
@@ -2,14 +2,14 @@ using GoatShooooting.Runtime;
 
 namespace GoatShooooting.Framework;
 
-public readonly record struct RunCompletion(int Score, bool Cleared);
+public readonly record struct RunCompletion(long Score, bool Cleared);
 
 public sealed class RunCompletionTracker
 {
     private SimulationStatus _previousStatus = SimulationStatus.Running;
     private bool _recordedCurrentRun;
 
-    public RunCompletion? Observe(SimulationStatus status, int score)
+    public RunCompletion? Observe(SimulationStatus status, long score)
     {
         if (status == SimulationStatus.Running)
         {
