+++ src/goat-shooooting.Runtime/SimulationTelemetry.cs
@@ -57,7 +57,7 @@ public sealed class SimulationTelemetry
     public int EnemyBulletsCleared { get; internal set; }
     public int EnemiesKilled { get; internal set; }
     public int BossesKilled { get; internal set; }
-    public int Score { get; internal set; }
+    public long Score { get; internal set; }
     public int ItemsSpawned { get; internal set; }
     public int ItemsCollected { get; internal set; }
     public int PlayerDeaths { get; internal set; }
